requirements.txt
# AI Brain
groq==1.0.0

# Text-to-Speech
pyttsx3==2.90
pyaudio==0.2.13

# Speech Recognition
SpeechRecognition==3.14.5

# Environment variables
python-dotenv==1.0.1

# Web UI
streamlit==1.29.0
rich==13.9.4

# Image processing & OCR
pillow==10.0.0
opencv-python==4.8.1.78
pytesseract==0.3.13

# Web & Video Search
duckduckgo-search==8.1.1
youtube-search-python==1.6.6

# Dependencies for TTS & other modules
numpy==1.25.2
importlib-metadata==6.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging>=23.0,<24  # allow Streamlit to work
protobuf==4.25.8
tenacity==8.5.0
tzlocal==5.3.1
validators==0.35.0






